Fixed compile warning.

+++ apps/app_zapras.c
@@ -160,7 +160,7 @@ static void run_ras(struct ast_channel *chan, char *args)
 			if (!res) {
 				/* Check for hangup */
 				if (chan->_softhangup && !signalled) {
-					ast_log(LOG_DEBUG, "Channel '%s' hungup.  Signalling RAS at %d to die...\n", chan->name, pid);
+					ast_log(LOG_DEBUG, "Channel '%s' hungup.  Signalling RAS at %d to die...\n", chan->name, (int)pid);
 					kill(pid, SIGTERM);
 					signalled=1;
 				}
